refactor #334: create components folder
