+++ components/molecules/table/components/tableData/components/battleTag/__tests__/index.tsx
