+++ components/molecules/table/components/tableData/components/battleTag/index.tsx
