+++ components/molecules/table/components/tableData/components/battleTag/styled.ts
