+++ components/molecules/table/components/tableData/components/challenges/__tests__/index.tsx
