+++ components/molecules/table/components/tableData/components/challenges/styled.ts
