+++ components/molecules/table/components/tableData/components/difficulty/index.tsx
