+++ components/molecules/table/components/tableData/components/difficulty/styled.ts
