+++ components/molecules/table/components/tableData/components/playersList/index.tsx
