+++ components/molecules/table/components/tableData/components/ratio/index.tsx
