+++ components/molecules/table/components/tableData/components/ratio/styled.ts
